requirements.txt
openai
pymupdf
PyPDF2
python-dotenv
tiktoken
pyyaml
